+++ pkg/provider/acceptance_source_postgres_test.go
@@ -32,6 +32,7 @@ func TestAccSourcePostgres_basic(t *testing.T) {
 					resource.TestCheckResourceAttr("materialize_source_postgres.test", "schema_name", "public"),
 					resource.TestCheckResourceAttr("materialize_source_postgres.test", "qualified_sql_name", fmt.Sprintf(`"materialize"."public"."%s"`, sourceName)),
 					resource.TestCheckResourceAttr("materialize_source_postgres.test", "size", "1"),
+					resource.TestCheckResourceAttr("materialize_source_postgres.test", "text_columns.#", "1"),
 					resource.TestCheckResourceAttr("materialize_source_postgres.test", "table.#", "2"),
 					resource.TestCheckResourceAttr("materialize_source_postgres.test", "table.0.name", "table1"),
 					resource.TestCheckResourceAttr("materialize_source_postgres.test", "table.0.alias", fmt.Sprintf(`%s_table1`, connName)),
@@ -70,7 +71,7 @@ func TestAccSourcePostgres_update(t *testing.T) {
 				Config: testAccSourcePostgresResource(roleName, secretName, connName, sourceName, source2Name, "mz_system"),
 			},
 			{
-				Config: testAccSourcePostgresResource(roleName, secretName, connName, newSourceName, source2Name, roleName),
+				Config: testAccSourcePostgresResourceUpdate(roleName, secretName, connName, newSourceName, source2Name, roleName),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckSourcePostgresExists("materialize_source_postgres.test"),
 					resource.TestCheckResourceAttr("materialize_source_postgres.test", "name", newSourceName),
@@ -78,14 +79,24 @@ func TestAccSourcePostgres_update(t *testing.T) {
 					resource.TestCheckResourceAttr("materialize_source_postgres.test", "schema_name", "public"),
 					resource.TestCheckResourceAttr("materialize_source_postgres.test", "qualified_sql_name", fmt.Sprintf(`"materialize"."public"."%s"`, newSourceName)),
 					resource.TestCheckResourceAttr("materialize_source_postgres.test", "size", "1"),
+					resource.TestCheckResourceAttr("materialize_source_postgres.test", "text_columns.#", "2"),
 					resource.TestCheckResourceAttr("materialize_source_postgres.test", "table.#", "2"),
 					resource.TestCheckResourceAttr("materialize_source_postgres.test", "table.0.name", "table1"),
 					resource.TestCheckResourceAttr("materialize_source_postgres.test", "table.0.alias", fmt.Sprintf(`%s_table1`, connName)),
-					resource.TestCheckResourceAttr("materialize_source_postgres.test", "table.1.name", "table2"),
-					resource.TestCheckResourceAttr("materialize_source_postgres.test", "table.1.alias", fmt.Sprintf(`%s_table2`, connName)),
+					resource.TestCheckResourceAttr("materialize_source_postgres.test", "table.1.name", "table3"),
+					resource.TestCheckResourceAttr("materialize_source_postgres.test", "table.1.alias", fmt.Sprintf(`%s_table3`, connName)),
 					resource.TestCheckResourceAttr("materialize_source_postgres.test", "publication", "mz_source"),
 					testAccCheckSourcePostgresExists("materialize_source_postgres.test_role"),
 					resource.TestCheckResourceAttr("materialize_source_postgres.test_role", "ownership_role", roleName),
+				),
+			},
+			{
+				Config: testAccSourcePostgresResource(roleName, secretName, connName, newSourceName, source2Name, roleName),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckSourcePostgresExists("materialize_source_postgres.test"),
+					resource.TestCheckResourceAttr("materialize_source_postgres.test", "table.#", "2"),
+					resource.TestCheckResourceAttr("materialize_source_postgres.test", "table.1.name", "table2"),
+					resource.TestCheckResourceAttr("materialize_source_postgres.test", "table.1.alias", fmt.Sprintf(`%s_table2`, connName)),
 				),
 			},
 		},
@@ -127,6 +138,74 @@ resource "materialize_role" "test" {
 }
 
 resource "materialize_secret" "postgres_password" {
+	name  = "%[2]s"
+	value = "c2VjcmV0Cg=="
+}
+
+resource "materialize_connection_postgres" "test" {
+	name = "%[3]s"
+	host = "postgres"
+	port = 5432
+	user {
+		text = "postgres"
+	}
+	password {
+		name          = materialize_secret.postgres_password.name
+		schema_name   = materialize_secret.postgres_password.schema_name
+		database_name = materialize_secret.postgres_password.database_name
+	}
+	database = "postgres"
+}
+
+resource "materialize_source_postgres" "test" {
+	name = "%[4]s"
+	postgres_connection {
+		name = materialize_connection_postgres.test.name
+	}
+
+	size  = "1"
+	publication = "mz_source"
+	table {
+		name  = "table1"
+		alias = "%[3]s_table1"
+	}
+	table {
+		name  = "table2"
+		alias = "%[3]s_table2"
+	}
+	text_columns = ["table1.id"]
+}
+
+resource "materialize_source_postgres" "test_role" {
+	name = "%[5]s"
+	postgres_connection {
+		name = materialize_connection_postgres.test.name
+	}
+
+	size  = "1"
+	publication = "mz_source"
+	table {
+		name  = "table1"
+		alias = "%[3]s_table_role_1"
+	}
+	table {
+		name  = "table2"
+		alias = "%[3]s_table_role_2"
+	}
+	ownership_role = "%[6]s"
+
+	depends_on = [materialize_role.test]
+}
+`, roleName, secretName, connName, sourceName, source2Name, sourceOwner)
+}
+
+func testAccSourcePostgresResourceUpdate(roleName, secretName, connName, sourceName, source2Name, sourceOwner string) string {
+	return fmt.Sprintf(`
+resource "materialize_role" "test" {
+	name = "%[1]s"
+}
+
+resource "materialize_secret" "postgres_password" {
 	name          = "%[2]s"
 	value         = "c2VjcmV0Cg=="
 }
@@ -155,13 +234,14 @@ resource "materialize_source_postgres" "test" {
 	size  = "1"
 	publication = "mz_source"
 	table {
-	  name  = "table1"
-	  alias = "%[3]s_table1"
+		name  = "table1"
+		alias = "%[3]s_table1"
 	}
 	table {
-	  name  = "table2"
-	  alias = "%[3]s_table2"
+		name  = "table3"
+		alias = "%[3]s_table3"
 	}
+	text_columns = ["table1.id", "table3.id"]
 }
 
 resource "materialize_source_postgres" "test_role" {
@@ -175,11 +255,11 @@ resource "materialize_source_postgres" "test_role" {
 	table {
 		name  = "table1"
 		alias = "%[3]s_table_role_1"
-	  }
-	  table {
+	}
+	table {
 		name  = "table2"
 		alias = "%[3]s_table_role_2"
-	  }
+	}
 	ownership_role = "%[6]s"
 
 	depends_on = [materialize_role.test]